:sparkles: feat: 로그인 구현완료, refresh token 진행 시작.

+++ src/main/java/com/umc/jaetteoli/domain/web/seller/dto/PostLoginRes.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class PostLoginRes {
     private TokenDto token;
-    private int sellerIdx;
+    private Long sellerIdx;
     private String name;
     private int first_login;
     private int menu_register;
